+++ PhysicalTrack/PhysicalTrack/Feature/Ranking/Ranking/RankingFeatrue.swift
@@ -49,13 +49,11 @@ struct RankingFeature {
         Reduce { state , action in
             switch action {
             case .onAppear:
-                state.consistency = .stubs //TODO: 꾸준함 랭킹 삭제
-                state.consistencyTop3 = state.consistency.prefix(3).map { $0 }
+                
                 return .merge(
-                    //TODO: 꾸준함 랭킹 나오면 반영
-//                    .run { [state = state] send in
-//                        await send(.consistencyRankingResponse(Result { try await rankingClient.fetchConsistency(state.accessToken)}))
-//                    },
+                    .run { [state = state] send in
+                        await send(.consistencyRankingResponse(Result { try await rankingClient.fetchConsistency(state.accessToken)}))
+                    },
                     .run { [state = state] send in
                         await send(.pushUpRankingResponse(Result { try await rankingClient.fetchPushUp(state.accessToken)}))
                     }
