+++ PhysicalTrack/PhysicalTrack/Feature/Ranking/RankingDetail/RankingDetailView.swift
@@ -70,7 +70,7 @@ struct ConsistencyRankingView: View {
                             
                             Spacer()
                             
-                            Text("\(data.daysActive)일째 운동 중")
+                            Text("\(data.streakCount)일째 운동 중")
                                 .foregroundStyle(.ptLightGray01)
                         }
                         .padding(12)
